[POI] filter entity type by favoredEntity
* add EntityType
* separate tests
