+++ skills/csharp/shared/SkillServiceLibrary/Models/AzureMaps/EntityType.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SkillServiceLibrary.Models.AzureMaps
+{
+    public static class EntityType
+    {
+        public static readonly string Country = "Country";
+
+        public static readonly string CountrySecondarySubdivision = "CountrySecondarySubdivision";
+
+        public static readonly string CountrySubdivision = "CountrySubdivision";
+
+        public static readonly string CountryTertiarySubdivision = "CountryTertiarySubdivision";
+
+        public static readonly string Municipality = "Municipality";
+
+        public static readonly string MunicipalitySubdivision = "MunicipalitySubdivision";
+
+        public static readonly string Neighbourhood = "Neighbourhood";
+
+        public static readonly string PostalCodeArea = "PostalCodeArea";
+    }
+}
