+++ skills/csharp/tests/pointofinterestskill.tests/Flow/PointOfInterestDialogTests.cs
@@ -243,223 +243,5 @@ namespace PointOfInterestSkill.Tests.Flow
                 .AssertReply(CheckForEvent())
                 .StartTestAsync();
         }
-
-        /// <summary>
-        /// Find parking nearby.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [TestMethod]
-        public async Task ParkingNearbyTest()
-        {
-            await GetTestFlow()
-                .Send(string.Empty)
-                .AssertReplyOneOf(this.ParseReplies(POIMainResponses.PointOfInterestWelcomeMessage))
-                .Send(BaseTestUtterances.LocationEvent)
-                .Send(FindParkingUtterances.FindParkingNearby)
-                .AssertReply(AssertContains(POISharedResponses.MultipleLocationsFound, new string[] { CardStrings.Overview }))
-                .Send(BaseTestUtterances.OptionOne)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Details }))
-                .Send(BaseTestUtterances.ShowDirections)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Route }))
-                .Send(BaseTestUtterances.StartNavigation)
-                .AssertReply(CheckForEvent())
-                .StartTestAsync();
-        }
-
-        /// <summary>
-        /// Find nearest parking.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [TestMethod]
-        public async Task ParkingNearestTest()
-        {
-            await GetTestFlow()
-                .Send(string.Empty)
-                .AssertReplyOneOf(this.ParseReplies(POIMainResponses.PointOfInterestWelcomeMessage))
-                .Send(BaseTestUtterances.LocationEvent)
-                .Send(FindParkingUtterances.FindParkingNearest)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Details }))
-                .Send(BaseTestUtterances.ShowDirections)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Route }))
-                .Send(BaseTestUtterances.StartNavigation)
-                .AssertReply(CheckForEvent())
-                .StartTestAsync();
-        }
-
-        /// <summary>
-        /// Find nearest parking and cancel.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [TestMethod]
-        public async Task ParkingNearestAndCancelTest()
-        {
-            await GetTestFlow()
-                .Send(string.Empty)
-                .AssertReplyOneOf(this.ParseReplies(POIMainResponses.PointOfInterestWelcomeMessage))
-                .Send(BaseTestUtterances.LocationEvent)
-                .Send(FindParkingUtterances.FindParkingNearest)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Details }))
-                .Send(GeneralTestUtterances.Cancel)
-                .AssertReply(AssertContains(POISharedResponses.CancellingMessage, null))
-                .StartTestAsync();
-        }
-
-        /// <summary>
-        /// Find nearest parking and help.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [TestMethod]
-        public async Task ParkingNearestAndHelpTest()
-        {
-            await GetTestFlow()
-                .Send(string.Empty)
-                .AssertReplyOneOf(this.ParseReplies(POIMainResponses.PointOfInterestWelcomeMessage))
-                .Send(BaseTestUtterances.LocationEvent)
-                .Send(FindParkingUtterances.FindParkingNearest)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Details }))
-                .Send(GeneralTestUtterances.Help)
-                .AssertReply(AssertContains(POIMainResponses.HelpMessage, null))
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Details }))
-                .Send(BaseTestUtterances.ShowDirections)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Route }))
-                .Send(BaseTestUtterances.StartNavigation)
-                .AssertReply(CheckForEvent())
-                .StartTestAsync();
-        }
-
-        /// <summary>
-        /// Find parking nearby.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [TestMethod]
-        public async Task PromptForCurrentAndParkingNearAddressTest()
-        {
-            await GetTestFlow()
-                .Send(string.Empty)
-                .AssertReplyOneOf(this.ParseReplies(POIMainResponses.PointOfInterestWelcomeMessage))
-                .Send(FindParkingUtterances.FindParkingNearAddress)
-                .AssertReply(AssertContains(POISharedResponses.PromptForCurrentLocation, null))
-                .Send(ContextStrings.Ave)
-                .AssertReply(AssertContains(POISharedResponses.CurrentLocationMultipleSelection, new string[] { CardStrings.Overview }))
-                .Send(BaseTestUtterances.OptionOne)
-                .AssertReply(AssertContains(POISharedResponses.MultipleLocationsFound, new string[] { CardStrings.Overview }))
-                .Send(BaseTestUtterances.OptionOne)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Details }))
-                .Send(BaseTestUtterances.ShowDirections)
-                .AssertReply(AssertContains(null, new string[] { CardStrings.Route }))
-                .Send(BaseTestUtterances.StartNavigation)
-                .AssertReply(CheckForEvent())
-                .StartTestAsync();
-        }
-
-        /// <summary>
-        /// Get directions.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [TestMethod]
-        public async Task GetDirectionsTest()
-        {
-            await GetTestFlow()
-                .Send(string.Empty)
-                .AssertReplyOneOf(this.ParseReplies(POIMainResponses.PointOfInterestWelcomeMessage))
-                .Send(BaseTestUtterances.LocationEvent)
-                .Send(RouteFromXToYUtterances.GetToMicrosoft)
-                .AssertReply(AssertContains(POISharedResponses.MultipleLocationsFound, new string[] { CardStrings.Overview }))
-                .Send(BaseTestUtterances.OptionOne)
-                .AssertReply(CheckForEvent())
-                .StartTestAsync();
-        }
-
-        /// <summary>
-        /// Get directions to some nearest place.
-        /// </summary>
-        /// <returns>A <see cref="Task"/> representing the asynchronous unit test.</returns>
-        [TestMethod]
-        public async Task GetDirectionsToNearestTest()
-        {
-            await GetTestFlow()
-                .Send(string.Empty)
-                .AssertReplyOneOf(this.ParseReplies(POIMainResponses.PointOfInterestWelcomeMessage))
-                .Send(BaseTestUtterances.LocationEvent)
-                .Send(RouteFromXToYUtterances.GetToNearestPharmacy)
-                .AssertReply(CheckForEvent())
-                .StartTestAsync();
-        }
-
-        private Action<IActivity> AssertStartsWith(string response, IList<string> cardIds)
-        {
-            return activity =>
-            {
-                var messageActivity = activity.AsMessageActivity();
-
-                if (response == null)
-                {
-                    Assert.IsTrue(string.IsNullOrEmpty(messageActivity.Text));
-                }
-                else
-                {
-                    Assert.IsTrue(ParseReplies(response, new StringDictionary()).Any((reply) =>
-                    {
-                        return messageActivity.Text.StartsWith(reply);
-                    }));
-                }
-
-                AssertSameId(messageActivity, cardIds);
-            };
-        }
-
-        private Action<IActivity> AssertContains(string response, IList<string> cardIds)
-        {
-            return activity =>
-            {
-                var messageActivity = activity.AsMessageActivity();
-
-                if (response == null)
-                {
-                    Assert.IsTrue(string.IsNullOrEmpty(messageActivity.Text));
-                }
-                else
-                {
-                    CollectionAssert.Contains(ParseReplies(response, new StringDictionary()), messageActivity.Text);
-                }
-
-                AssertSameId(messageActivity, cardIds);
-            };
-        }
-
-        private void AssertSameId(IMessageActivity activity, IList<string> cardIds = null)
-        {
-            if (cardIds == null)
-            {
-                return;
-            }
-
-            for (int i = 0; i < cardIds.Count; ++i)
-            {
-                var card = activity.Attachments[i].Content as JObject;
-                Assert.AreEqual(card["id"], cardIds[i]);
-            }
-        }
-
-        /// <summary>
-        /// Asserts bot response of Event Activity.
-        /// </summary>
-        /// <returns>Returns an Action with IActivity object.</returns>
-        private Action<IActivity> CheckForEvent(PointOfInterestDialogBase.OpenDefaultAppType openDefaultAppType = PointOfInterestDialogBase.OpenDefaultAppType.Map)
-        {
-            return activity =>
-            {
-                var eventReceived = activity.AsEventActivity()?.Value as OpenDefaultApp;
-                Assert.IsNotNull(eventReceived, "Activity received is not an Event as expected");
-                if (openDefaultAppType == PointOfInterestDialogBase.OpenDefaultAppType.Map)
-                {
-                    Assert.IsFalse(string.IsNullOrEmpty(eventReceived.MapsUri));
-                }
-                else if (openDefaultAppType == PointOfInterestDialogBase.OpenDefaultAppType.Telephone)
-                {
-                    Assert.IsFalse(string.IsNullOrEmpty(eventReceived.TelephoneUri));
-                }
-            };
-        }
     }
 }
